LPS-99657 Class that implements ModelResourcePermission should be in package *.security.permission.resource
